update layout files names
